requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-environ==0.10.0
idna==3.4
mega.py @ git+https://github.com/pythonbrad/mega.py@cb5914f72a3829c467b0dd0e0a2cbfa48e82e954
mysqlclient==2.1.1
pathlib==1.0.1
Pillow==9.5.0
pycryptodome==3.17
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tenacity==5.1.5
urllib3==1.26.15
